requirements.txt
affine==2.4.0
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contextily==1.4.0
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.104.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
fiona==1.9.5
fonttools==4.43.1
geographiclib==2.0
geopandas==0.14.0
geopy==2.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
httpx-oauth==0.13.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
makefun==1.15.1
Mako==1.2.4
mapped==20.4.2
MarkupSafe==2.1.3
matplotlib==3.8.0
mercantile==1.2.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
passlib==1.7.4
Pillow==10.1.0
plotly==5.18.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
rasterio==1.3.9
requests==2.31.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
xyzservices==2023.10.1
